requirements.txt
# Core dependencies
playwright>=1.40.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
torch>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
numpy>=1.24.0
transformers>=4.30.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

